fix: resolve foreign key data type conflicts and stabilize server

+++ models/RefreshToken.js
@@ -3,7 +3,7 @@ const sequelize = require('../config/database');
 
 const RefreshToken = sequelize.define('RefreshToken', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  userid: { type: DataTypes.BIGINT, allowNull: false },
+  userid: { type: DataTypes.INTEGER, allowNull: false },
   token: { type: DataTypes.TEXT, allowNull: false },
   expiry_date: { type: DataTypes.DATE, allowNull: false }
 }, {
